Refactor package structure and modify resource paths for consistency
Update all Java package names from `de.greluc.sc.sckillmonitor` to `de.greluc.sc.sckm` for brevity and alignment. Adjust corresponding FXML resource paths and module declarations to reflect the new structure. This change improves code readability and simplifies package hierarchy management.
