+++ src/main/java/de/greluc/sc/sckm/data/package-info.java
@@ -23,4 +23,4 @@
  * @since 1.0.0
  * @version 1.0.0
  */
-package de.greluc.sc.sckillmonitor;
+package de.greluc.sc.sckm.data;
