--- src/main/resources/de/greluc/sc/sckillmonitor/SettingsView.fxml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.RowConstraints?>
-
-<!--~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  ~ SC Kill Monitor                                                                               ~
-  ~ Copyright (C) 2025-2025 SC Kill Monitor Team                                                  ~
-  ~                                                                                               ~
-  ~ This file is part of SC Kill Monitor.                                                         ~
-  ~                                                                                               ~
-  ~ SC Kill Monitor is free software: you can redistribute it and/or modify                       ~
-  ~ it under the terms of the GNU General Public License as published by                          ~
-  ~ the Free Software Foundation, either version 3 of the License, or                             ~
-  ~ (at your option) any later version.                                                           ~
-  ~                                                                                               ~
-  ~ SC Kill Monitor is distributed in the hope that it will be useful,                            ~
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of                                ~
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the                                 ~
-  ~ GNU General Public License for more details.                                                  ~
-  ~                                                                                               ~
-  ~ You should have received a copy of the GNU General Public License                             ~
-  ~ along with SC Kill Monitor. If not, see <http://www.gnu.org/licenses/>.                       ~
-  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~-->
-
-<GridPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="300.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="de.greluc.sc.sckillmonitor.controller.SettingsViewController">
-  <columnConstraints>
-      <ColumnConstraints hgrow="NEVER" maxWidth="10.0" minWidth="10.0" prefWidth="10.0" />
-    <ColumnConstraints hgrow="SOMETIMES" maxWidth="155.0" minWidth="10.0" prefWidth="155.0" />
-    <ColumnConstraints hgrow="ALWAYS" minWidth="10.0" prefWidth="100.0" />
-      <ColumnConstraints hgrow="NEVER" maxWidth="10.0" minWidth="10.0" prefWidth="10.0" />
-  </columnConstraints>
-  <rowConstraints>
-      <RowConstraints maxHeight="10.0" minHeight="10.0" prefHeight="10.0" vgrow="NEVER" />
-    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-      <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-      <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-      <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-      <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-      <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-      <RowConstraints maxHeight="10.0" minHeight="10.0" prefHeight="10.0" vgrow="NEVER" />
-  </rowConstraints>
-   <children>
-      <Button mnemonicParsing="false" onAction="#onSave" text="Save" GridPane.columnIndex="1" GridPane.rowIndex="7" />
-      <Label text="Path to LIVE:" GridPane.columnIndex="1" GridPane.rowIndex="1" />
-      <TextField fx:id="inputPathPtu" onMouseClicked="#onPtuClicked" GridPane.columnIndex="2" GridPane.rowIndex="2" />
-      <Label text="Path to PTU:" GridPane.columnIndex="1" GridPane.rowIndex="2" />
-      <Label text="Path to EPTU:" GridPane.columnIndex="1" GridPane.rowIndex="3" />
-      <Label text="Path to HOTFIX:" GridPane.columnIndex="1" GridPane.rowIndex="4" />
-      <Label text="Path to TECH-PREVIEW:" GridPane.columnIndex="1" GridPane.rowIndex="5" />
-      <TextField fx:id="inputPathLive" onMouseClicked="#onLiveClicked" GridPane.columnIndex="2" GridPane.rowIndex="1" />
-      <TextField fx:id="inputPathEptu" onMouseClicked="#onEptuClicked" GridPane.columnIndex="2" GridPane.rowIndex="3" />
-      <TextField fx:id="inputPathHotfix" onMouseClicked="#onHotfixClicked" GridPane.columnIndex="2" GridPane.rowIndex="4" />
-      <TextField fx:id="inputPathTechPreview" onMouseClicked="#onTechPreviewClicked" GridPane.columnIndex="2" GridPane.rowIndex="5" />
-      <Label text="Path to Custom Log:" GridPane.columnIndex="1" GridPane.rowIndex="6" />
-      <TextField fx:id="inputPathCustom" onMouseClicked="#onCustomClicked" GridPane.columnIndex="2" GridPane.rowIndex="6" />
-   </children>
-</GridPane>
